requirements.txt
anyio==3.6.2
asgiref==3.5.2
azure-core==1.26.1
azure-monitor-opentelemetry-exporter==1.0.0b10
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
fastapi==0.88.0
fixedint==0.1.6
flake8==6.0.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
isodate==0.6.1
mccabe==0.7.0
msrest==0.7.1
mypy-extensions==0.4.3
oauthlib==3.2.2
opentelemetry-api==1.14.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-asgi==0.35b0
opentelemetry-instrumentation-fastapi==0.35b0
opentelemetry-instrumentation-logging==0.35b0
opentelemetry-instrumentation-sqlalchemy==0.35b0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
wrapt==1.14.1
